frontend/flyout/files: extract header into its own component
